requirements.txt
pandas~=1.5.3
fastapi~=0.92.0
python-multipart~=0.0.6
python-dotenv~=1.0.0
starlette~=0.25.0
pydantic~=1.10.5
uvicorn~=0.20.0